Explain why something I got worried about today isn't possible.
If, in a read-write tree, it were possible for instantiate to be
called with ref->ft_info.del NULL and a different page_del (taken from
the on-disk cell) passed in, there'd be a problem in parent-page
reconciliation: if a checkpoint needs to write the original page, the
code path to fix last week's bug would fire mistakenly and then see a
NULL modify->page_del and conclude, perhaps wrongly, that the deletion
is globally visible. However, this can't happen. If the page state is
WT_REF_DELETED, ref->ft_info.del is always the current deletion
information; it's only for pages in WT_REF_DISK state that are
actually deleted that a different page_del is passed in. And that only
happens in readonly trees, because it's a consequence of eviction
discarding instantiated but officially unmodified pages.

So, false alarm.