test_reconciliation_tracking: remove null-args test
Unfortunately, this test is less valid than I thought. Not sure how it
was happening so reliably, but the null-dereference was getting
optimised out on Linux builds. Windows (and non-optimised Linux
builds) were being a little less clever, and segfaulting.

Arguably, __ovfl_discard_verbose should be fixed to deal with null
arguments, but it's a static function with only a couple of users,
none of which can call this with a null page pointer.