It is possible (and legal) for modify->page_del to go away before abort.
modify->instantiated and modify->page_del are cleared if the
transaction aborts, but it's also cleared at the first reconciliation
of the page after instantiation. So don't assert that
modify->instantiated is still set.

Note that the reason it's there is that internal page reconciliation
needs it in the case where the truncate was committed before a
checkpoint starts but the instantiation (and thus the creation of
page->modify) happened after. By default reconciliation assumes that
because the modify structure was created after the checkpoint started
that nothing in it can be needed in the checkpoint.