Bail on invisible truncates in __rec_child_deleted during eviction.
For checkpoints we leave the page dirty because we need to write
something. For eviction, we can just evict something else instead, and
unlike for leaf pages, there's not much gained by regenerating the
page image of an internal page if we need to keep it in memory.

Otherwise we hit an assertion in __rec_write_page_status that expects
WT_REC_HS to be set. We don't set WT_REC_HS on internal pages, because
it doesn't make sense. While maybe the principled way to deal with
this is to add another flag comparable to WT_REC_HS for internal
pages, which would let the eviction proceed, this doesn't seem
worthwhile.