Allow using the checkpoint-cursor txn to read the metadata.
Adjust the visibility code to test the flag in transaction for being a
checkpoint-cursor transaction rather than test the tree for being a
checkpoint, so we use the checkpoint-cursor visibility logic even if
we're reading a non-checkpoint tree. This prevents it from returning
mutually inconsistent results.

It might still return wrong results if asked about specific
transaction ids (since the snapshot might be from a different run) but
since the metadata is read-uncommitted this shouldn't come up.

(Plus, if it didn't work we'd probably already know it, since it's
been doing this since the original checkpoint cursors merge and that's
some months back now.)