Revert the changes for setting certain list config defaults to "empty".
Things break if the value recorded in config_def.c is "" (that is,
verbose=,other-stuff) rather than "[]" (that is, verbose=[],other-stuff).

It appears that somehow the default session verbose setting ends up on
disk and then wt dump gets entangled with it if it's fully empty
rather than an empty list, because it gets passed to session_create as
the user config and that's subjected to a stronger schema check than
the default config ever is. Or something like that.

Correct solution is probably to write [] into config_def.c for all
empty lists.

Correct solution also is probably to fix whatever causes session
config to be written to disk, as that seems like a bug.