WT-9500 Fix RTS to use cell time window instead of key/value timestamps
RTS reads the timestamps present as part of the key/value of a history store
record to decide it's visibility. These timestamp values can be wrong when
there exists a globally visible tombstone in the update list and leads to
removal of the entire key by restoring a wrong update/tombstone combination.

To avoid RTS generating wrong update/tombstones, always read the cell time
window information to decide the HS update visibility.