WT-9500 Fix RTS to use cell time window instead of key/value timestamps (#8073)
RTS reads the timestamps present as part of the key/value of a history store
record to decide its visibility. These timestamp values can be wrong when
there exists a globally visible tombstone in the update list and leads to
the removal of the entire key by restoring a wrong update/tombstone combination.

To avoid RTS generating wrong updates/tombstones, always read the cell time
window information to decide the HS update visibility.