Implement fast-truncate for VLCS.
Mostly it's possible to just discard pages; reuse of the discarded
chunks of namespace gets directed to the append list of the next leaf
page to the left in the tree.

The exception is that one must avoid discarding the leftmost child of
a VLCS internal page, because the internal page structure prevents
this namespace from shifting any further to the left. Instead one ends
up with a gap between the internal page's start recno and the first
child's start recno, and this leads to trouble.

Take the following steps to avoid this:
 - Insert an empty deleted leaf at internal page inmem time if the
first child begins after the internal page itself.
 - Don't allow eviction to trigger a reverse split going upwards from
the leftmost child, as that discards the page. (In most cases the same
reverse split will then promptly happen anyway, coming from the next
child.)
 - Don't discard the leftmost child when doing a split, even if it's
deleted. (In effect, pretend in this case that the deletion isn't
globably visible yet.)

Also:
 - Add deleted-page instantiation code for VLCS.
 - Add support for deleted address cells to column internal pages
   (both in inmem and in reconciliation).
 - Allow checkpoint cleanup to discard pages from VLCS trees.
 - Cause VLCS leaf pages that reconcile with a single deleted-address
   cell to reconcile as empty instead.
 - Reset mod_col_split_recno after leaf reconciliation because we now
   might end up appending to the page and splitting again later.
 - Adjust verify to tolerate gaps in the recno space, and allow
   the leftmost child to be empty with no address.
 - Fix places in the system that assume only row-store does
   fast-truncate.
 - Update comments.
 - Teach s_void about WT_PAGE_ALLOC_AND_SWAP.
 - Update various tests.
 - Add two new tests; one tests reading in namespace gaps, the other
   makes sure nothing blows up (or takes degenerately long) with very
   large namespace gaps.

Note that fast-truncate for FLCS has a whole additional set of
problems because FLCS has to iterate through (rather than skip over)
gaps in the namespace. Currently this doesn't looks feasible to
support.