Make FLCS test_salvage01 more robust.
When looking for the intended place for intentional data corruption,
search for a sequence of bytes in adjacent rows instead of just one
byte.

This can still fail if the sequence ends up sitting across a page
boundary, but it's much less likely to get false matches in bad places
that result in test failure than looking for just one byte.