Evict any committed truncate, not just globally-visible ones.
This turns out to be more than a one-line change because for eviction
threads the check happens before we get the snapshot that we'll be
using to check for visibility. So in that case we want to check
explicitly if the truncate's committed. Otherwise, if we have a
snapshot, use it; otherwise fall back to a global visibility check.

Include the test Hari posted in the ticket as test_truncate15, with a
few adjustments. Most notably, the cache overflow condition now
results in reads failing with WT_ROLLBACK, so catch that and fail if
we see it.

(Otherwise the test harness retries the whole test, and that isn't
what we want.)

Also, I had to make it larger to get it to fail reliably.

Note 1: the transition from page_del_active to page_del_visible is not a
functional change; page_del_active() returns !page_del_visible() and
is going to go away.

Note 2: test_truncate13/14 are pending in WT-9603.