Stop using mod_col_split_recno for VLCS.
This is somewhat provisional but I think it's the right thing to do.

Basically it is used for two things: (a) to avoid creating a gap at
the end of a page if there's a gap between the last entry in the
append list and the next entry that was moved to another page by an
in-memory split; and (b) cross-check in col_modify that we didn't end
up on the wrong side of an in-memory split as a result of some sort of
race or other issue.

We no longer need to avoid creating such gaps (they're equivalent to
gaps created by truncate, though perhaps smaller) and the cross-check
is nice but not critical.

Meanwhile there's a problem I've been worried about. Suppose page A
splits into page A and page B, then page B is truncated and discarded,
then someone writes into B's namespace. That write will be sent to A.
And with the mod_col_split_recno logic in place, if this happens
before A is reconciled (when mod_col_split_recno gets cleared) it'll
cause a false positive in the check in col_modify. Disabling
mod_col_split_recno for VLCS avoids this problem. Meanwhile, there
doesn't seem to be any other good way to deal with it. And while it
seems to be more or less impossible to set up on purpose, there's also
no reason to think it can't ever happen...