WT-9393 Reading fast truncated pages in a readonly database can lead to cache stuck (#8176)
This second attempt also fixes the check rules. The truncate must be committed, even if visible.

Also add an assert in __rec_child_deleted: if something gets messed up
and we try to write out an uncommitted truncate, it should die there
and not wait until the output page fails verify. Since (among other
things) it might not in some cases.