WT-9603 Fast-truncate for VLCS (#8129)
* Implement fast-truncate for VLCS.

Mostly it's possible to just discard pages; reuse of the discarded
chunks of namespace gets directed to the append list of the next leaf
page to the left in the tree.

The exception is that one must avoid discarding the leftmost child of
a VLCS internal page, because the internal page structure prevents
this namespace from shifting any further to the left. Instead one ends
up with a gap between the internal page's start recno and the first
child's start recno, and this leads to trouble.

Take the following steps to avoid this:
 - Insert an empty deleted leaf at internal page inmem time if the
first child begins after the internal page itself.
 - Don't allow eviction to trigger a reverse split going upwards from
the leftmost child, as that discards the page. (In most cases the same
reverse split will then promptly happen anyway, coming from the next
child.)
 - Don't discard the leftmost child when doing a split, even if it's
deleted. (In effect, pretend in this case that the deletion isn't
globably visible yet.)

Also:
 - Add deleted-page instantiation code for VLCS.
 - Add support for deleted address cells to column internal pages
   (both in inmem and in reconciliation).
 - Allow checkpoint cleanup to discard pages from VLCS trees.
 - Cause VLCS leaf pages that reconcile with a single deleted-address
   cell to reconcile as empty instead.
 - Reset mod_col_split_recno after leaf reconciliation because we now
   might end up appending to the page and splitting again later.
 - Adjust verify to tolerate gaps in the recno space, and allow
   the leftmost child to be empty with no address.
 - Fix places in the system that assume only row-store does
   fast-truncate.
 - Update comments.
 - Teach s_void about WT_PAGE_ALLOC_AND_SWAP.
 - Update various tests.
 - Add two new tests; one tests reading in namespace gaps, the other
   makes sure nothing blows up (or takes degenerately long) with very
   large namespace gaps.

Note that fast-truncate for FLCS has a whole additional set of
problems because FLCS has to iterate through (rather than skip over)
gaps in the namespace. Currently this doesn't looks feasible to
support.

* The chewing gum fell off the logic in FLCS test_salvage01.

Stick it back on for now. Should really fix it properly one of these
days.

* Honor self.extraconfig in the new tests. (oops)

* Update/add/improve some comments as suggested.

* Some minor additional review changes.

* Add stats as requested.

* Take a shot at making __inmem_col_int tidier.

* Unpack the first cell to check if we need an extra ref slot.

This is a bit untidy but it's much less untidy than the previous code,
which was always allocating an extra ref and then discarding it if it
wasn't used.

Also probably faster. Unpacking isn't free, but it's certainly cheaper
than malloc.

* Stop using mod_col_split_recno for VLCS.

This is somewhat provisional but I think it's the right thing to do.

Basically it is used for two things: (a) to avoid creating a gap at
the end of a page if there's a gap between the last entry in the
append list and the next entry that was moved to another page by an
in-memory split; and (b) cross-check in col_modify that we didn't end
up on the wrong side of an in-memory split as a result of some sort of
race or other issue.

We no longer need to avoid creating such gaps (they're equivalent to
gaps created by truncate, though perhaps smaller) and the cross-check
is nice but not critical.

Meanwhile there's a problem I've been worried about. Suppose page A
splits into page A and page B, then page B is truncated and discarded,
then someone writes into B's namespace. That write will be sent to A.
And with the mod_col_split_recno logic in place, if this happens
before A is reconciled (when mod_col_split_recno gets cleared) it'll
cause a false positive in the check in col_modify. Disabling
mod_col_split_recno for VLCS avoids this problem. Meanwhile, there
doesn't seem to be any other good way to deal with it. And while it
seems to be more or less impossible to set up on purpose, there's also
no reason to think it can't ever happen...

* Create a macro for the page_del -> time aggregate op used in internal page reconciliation.

(This also tidies the way it works slightly.)

* Enable VLCS in test_truncate15.

* Fix s_function failure after merging WT-9645.

* Add the rest of the update for truncate15. (oops)