Use atomics to clear mod->instantiated and discard ref->page_del.
Otherwise we can free it twice if we hit two small race windows in
succession while one thread is finishing with parent page
reconciliation and the other aborts the transaction, in the fairly
unlikely circumstance that these happen together after the page is
instantiated.

Pretty sure this condition has been present on develop for a while
without triggering. :-|