Fix row-store assertion that appeared in testing.
It seems that the optimization to generate an empty page instead of
reading in a deleted page containing only obsolete values breaks
row-store verify.

This is because the key consistency check requires the physical
presence of the keys, regardless of whether their values are deleted.
Producing an empty page instead seems to be harmless in general
(though bad if the obsolete page we skip turns out for some other
reason to contain disordered keys) but the check is initialized from
the leftmost page in the tree and if _that_ one is replaced with an
empty page things go off the rails.

There doesn't seem to be any good solution for this; at least one of
the keys on that page is needed. So, work around it by skipping the
optimization when doing verify; and for consistency with other nearby
logic, also skip it for salvage and upgrade. (It seems like skipping
it for salvage might be a good idea, in any case.)

Add a python test that exercises this case. It turns out it's not so
easy to get a tree with a leftmost deleted page, but it is possible
with some mucking about.