Fix column-store assertion that appeared during testing.
In the WT-9603 changes I overlooked the possibility that an entirely
empty page (no cells, no append list entries at all, even ones with no
updates) might get reconciled, and if this happens the code that's
supposed to convert a page that's one deleted cell to an empty page
asserts. Skip that code if the page is already empty.

It is apparently possible for such an entirely empty page to get
marked dirty (not sure how) so the change in this changeset to create
empty pages instead of loading and instantiating pages full of
obsolete values can cause this assertion to trigger.

(Explicitly marking those empty pages dirty triggers the problem right
away, so I'm pretty sure that's what's happening.)

I don't have a python test for this one; can't figure out how to set
the situation up.