Fix very small race window between mod->instantiated and ref locking.
It is theoretically possible for mod->instantiated to transition from
true to false while we do WT_REF_CAS_STATE. In that case, unlock and
try again. Either the child just got reconciled (in which case we will
have a mod->rec_result and take a different path) or the truncation
got rolled back. Either way, mod->instantiated can't switch back to
true.