Fix the failure that appeared this past weekend.
When skipping deleted-page instantiation for verify, salvage, and
upgrade, still create page->modify and set page->modify->instantiated.

Otherwise we assert in eviction because the page has a page_del but
isn't marked instantiated.

It is not absolutely clear that this is actually necessary; without it
and without the assert, eviction will change the ref state to
WT_PAGE_DISK instead of WT_REF_DELETED and the fact of the deletion
will be lost; but in neither verify nor salvage do we particularly
care about that as long as the handle is closed at the end so the
in-memory state isn't retained. However, for upgrade it's less clear,
especially since upgrade doesn't currently do anything but might in
the future; and meanwhile, making the truncate code more robust means
there are probably other assertions that will fire if this is allowed
to happen.

If at some point we decide that allocating a modify structure just to
hold the instantiated flag is undesirable we could probably find a
different way to encode the information; for example, it could be
handled as an additional ref state. That kind of change doesn't belong
in this patchset though.