WT-9772 Fix RTS to abort the updates using the recovered checkpoint snapshot
All the running transactions are either committed/rollbacked before
RTS runs, so there is no need to check for the transaction to abort an
update. But with an unstable fast trucate page reads back into memory
and instantiate all the keys with a tombstone that needs to be aborted
according to the recovered checkpoint snapshot.