Add a fast-truncate page to the architecture guide.
Improve s_docs so we don't need non-words appearing in certain
filenames and symbols (e.g. "vrfy") to be spelling words. This is done
with ad hoc seddery because there are not many cases and not that many
new ones likely to appear.

Add some spelling words.

Drop most of the old brief description in arch-btree and refer to the
new page instead.

Also update the discussion of the transactionality of truncate
operations. I'm not changing what the user-facing docs say without
discussion, but we can afford to document the current understanding
here instead of having it buried in various tickets and github
comments.

(The known bug referred to is WT-4158. Didn't put the ticket number in
the docs because that doesn't seem like the right thing, but can
certainly add it if desired.)