WT-7833 wt_btree_switch_object needs to coordinate with concurrent write requests to btree (#8044)
* Allow concurrent writes to any number of WT_BLOCK handles. When we switch files there can
be any number of writers in flight to the previous file. Because checkpoints are limited to a
single file, all writers must have drained after a subsequent checkpoint, and then we can close
the original WT_BLOCK reference.

* Lock when extending the WT_BLOCK list of cached handles, two threads can race doing the
extension (but not when checking the cached handle list, of course).

* If you checkpoint the current object, then switch objects, and never checkpoint the new object,
you can end up with checkpoint drain chains at database close. It's slightly difficult to handle
because the objects won't be closed in any particular order, assert we're in close because any
other path would be a sign of a serious problem.

* Discard the previous active WT_BLOCK handles after each subsequent checkpoint.

* Removed "flush locks" in some test programs that prevented concurrent updates and flush_tier

Co-authored-by: Don Anderson <donald.anderson@mongodb.com>