WT-9792 Fix RTS to remove globally visible update of the data store update from the history store (#8240)
The globally visible start time window historical updates are missed
to remove from the history store when the same update is restored as
a stable update in the data store due to the start timestamp comparison.
For a globally visible update, the start timestamp is reset to 0.
Adjust the check to compare the stop timestamp instead of start to
remove the same stable update in the history store.

Along with the above fix, let the history store will be scanned at the
end of RTS for any unstable updates to remove in all scenarios as it used
to do only during recovery. Avoid performing the final pass in an in-memory
config.