WT-9763 Return EBUSY if the time window of inserting record does not match with history store time window (#8218) (#8246)
This bug was one of the cases where the data store and the history store may contain the same value. In this case, skip inserting the same value into the history store again. The fix is the time window of the inserting record and the history store record are compared to make sure that the same record is not being inserted again.

(cherry picked from commit f45ebf82dba6e0d88edccb6c3047a717669b4562)