WT-9457 Preserve ckpt_most_recent value across restart (#8245)
* WT-9457 Preserve ckpt_most_recent value across restart

The core change here is to initialize conn->ckpt_most_recent from the time on the last checkpoint in the metadata. This ensures that new checkpoint and flush times after (re-)starting WT will be greater than those used before we shutdown. The change is just a couple lines of code to load this value at the same time we load the base write generation.

The other changes here include:

New test_bug029.py that reproduces the original failure mode from WT-9457.

Changes to the names of various initialization functions in meta_ckpt.c. Previously, they just loaded the base write gen, and could have names like wt_metadata_init_base_write_gen. 

I've standardized the the parsing of the time value from the checkpoint config since it needs to be done outside the config parser.

I eliminated the code in ckpt_last_name() that returns a checkpoint time of zero if the checkpoint has a write generation before the current base. As I understand it, this is no longer necessary because this PR ensures that checkpoints with earlier write generations will also have smaller clock times.

I've changed the error handling in ckpt_load to eliminate the use of a special format: label in the code.