WT-9811 Fast-truncate page for the architecture guide (#8237)
Improve s_docs so we don't need non-words appearing in certain
filenames and symbols (e.g. "vrfy") to be spelling words. This is done
with ad hoc seddery because there are not many cases and not that many
new ones likely to appear.

Also update the discussion of the transactionality of truncate
operations. I'm not changing what the user-facing docs say without
discussion, but we can afford to document the current understanding
here instead of having it buried in various tickets and github
comments.