WT-9792 Fix RTS to remove globally visible update of the data store update from the history store (#8261)
The globally visible start time window historical updates are missed
to remove from the history store when the same update is restored as
a stable update in the data store due to the start timestamp comparison.
For a globally visible update, the start timestamp is reset to 0.
Adjust the check to compare the stop timestamp instead of the start to
remove the same stable update in the history store.

(cherry picked from commit f36a8e9d80dde82ceb11805e6437e12a2d784ae6)