WT-9742 Add timestamp support to Simple and Complex data sets in Python tests. (#8286)
* WT-9742 Add timestamp support to Simple and Complex data sets.
These data set classes are now outfitted to use an optional WiredTigerTimeStamp (timestamp allocator) object.
If a timestamp allocator is attached to the data set, then all modifications use the timestamp allocator
to get a commit timestamp, which also means that each individual modification is in an explicit transaction.
This is done via instrumented cursors that are used by the data sets.  Running with the tiered hook causes
a timestamp allocator to be used.

Any test that uses data sets, that want to use its own cursors, now need to use a *DataSet.open_cursor() call
to get an instrumented cursor.  Likewise, truncate calls on the data set need to be transactional, so use
a *DataSet.truncate() call.

There are some exceptions to this.  If a transaction is already in progress, we do not want to start a
new transaction, but rather use timestamp_transaction to give a timestamp to the existing transaction.

To make this all work, we now (in the tiered hook) override {begin,commit,rollback}_transaction so we can track
thr current transactional state on the session object.  And because our SWIG/Python interface may give us new
Python objects every time we reference WT_SESSION.connection, and new objects won't have that new transactional state,
we now create a different back reference to the connection from every session, and to the session from every
cursor.  Someday we may be able to fix the SWIG/Python layer to "do the right thing" with WT_SESSION.connection and
WT_CURSOR.session, but this is our workaround for now.  And to be clear, these new references (WT_SESSION._connection
and WT_CURSOR._session) are only available when running with the tiered hook.

With all this, most uses of the data sets work with timestamps.  There are some tests that do their own timestamping,
that will skip the tiered/timestamp mashup.  And there are about 5 tests that fail for still unknown reasons, we'll
tackle those in other tickets.