WT-9959 Remove statistic configs from test/format files. (#8341)
The statistic config was removed from test/format as we want it turned on by default and to run every time therefore the config was made redundant. The config's remnants in config_def.c and connection.h and also wts.c are removed in this PR so test/format builds properly.