WT-10030 Checkpoint cleanup remove the obsolete deleted pages
Checkpoint cleanup checks the obsoleteness of the deleted pages and
mark the parent page to remove it.

Also fail evicting of an internal page  when it cannot write the
fast truncate information to the disk.

Fix a bug where checkpoint skips an internal page when the fast truncate
operation is perfomred without a timestamp.