WT-10239 Implement dynamic table deletion in Workgen (#8579)
Using a reference counter when a table in Workgen is used.

Able to ask Workgen to prepare a table for deletion using the new Workgen API remove_table
When this is the case, the table cannot be used by any other thread. However, the threads currently using the table can finish their job.

A new API garbage_collection has been created so Workgen can delete the data associated with those tables when it is safe to do so. This API only deletes the data in Workgen, the table will still be on disk and the WiredTiger API session:drop should be called on the table.