WT-10336 Fix session stats bytes read increment with wrong info
Due to not initializing the session stat offset and its default value
maps to bytes read. Any other connection lock trackings will miskanely
update the session bytes read stat as they are not tracked specifically
in the session.

Initialize it explictly with -1 and check for the proper offset before
increment the session stats fixes the wrong increment of bytes read