WT-10575 Fix checkpoint cursor to return only visible stable data
Returning only visible data can lead to inconsisten results when
a prepared transaction gets committed in parallel to the checkpoint
with a commit timestamp less than or equal to the checkpoint stable
timestamp.

To avoid producing torn transaction results, along with the commit
timestamp visibility, the data must be durable according to the
checkpoint stable timestamp.