fix: specify IMPORTED_IMPLIB for builtin compressors on Windows
Add IMPORTED_IMPLIB property for imported SHARED wt::<compressor> libraries.
On Windows, find_library() always find a .lib(or .dll.a) file. The IMPORTED_IMPLIB
should be set to the .lib part of a DLL.
Since IMPORTED_LOCATION is needed on non-Windows and is optional on Windows and
IMPORTED_IMPLIB is optional on non-Windows and is needed on Windows, simply setting
both of them to ${HAVE_LIB<compressor>} will work.