Add keyblob.py tool and initial dump window support
Dump window support is untested as it is unclear how to traffic in a
numerical value via argv. The most likely function is util_str2num but
it works with u64 and the current wiindowing code takes advantage of a
signed value. Easy enough to downcast, but need to research prior art
first.