Migrate to use c11 stdatomic.h instead of gcc built ins
I have to disable building all the c++ code in wt as c11 is not compatible with c++
Simple tweaking of the gcc.h header file to provide a c++ version doesn't work either.
Significant work is needed to put in the c++ code and wt to be able to build with c11
standard header files.