WT-9812 Chunk cache (#8784)
The implementation of the chunk cache. The chunk cache caches fixed-sized chunks of data from object files. It is meant to be used in the case when the object is stored on a remote tier with a high access latency, and it is desirable to cache parts of the objects on a local storage. 
---------

Co-authored-by: Sasha Fedorova <sasha.fedorova@10gen.com>