WT-11031 Fix RTS to skip tables with no time window information in checkpoint
The versions before 4.4 only writes the checkpoint with no time window
information and in those versions only the stable data is written to the disk.
There is no need to verify these tables as part of the rollback to stable
as these tables contains only the stable data to avoid unnecessary verification.