WT-10820 Checkpoint to mark the tree dirty when the table has more data
Eviction can run in parallel to the checkpoint and can able to write
more data than the checkpoint can see and checkpoint operation
includes the changes from the eviction as part of the checkpoint
can lead to the scenario where this table is ignored in the next
checkpoint due to the table being clean leads to a scenario where
the log records of a table are replayed from the prior checkpoint
location than the current checkpoint as part of the backup restore.

To avoid ignoring these tables as part of the next checkpoint, mark
the table dirty whenever checkpoint detects that it has more data
than checkpoint can see.