WT-9457 Preserve ckpt_most_recent value across restart
The core change here is to initialize conn->ckpt_most_recent from the time on the last checkpoint in the metadata. This ensures that new checkpoint and flush times after (re-)starting WT will be greater than those used before we shutdown. The change is just a couple lines of code to load this value at the same time we load the base write generation.