WT-11064 Skip globally visible tombstones as part of update obsolete check
The globally visible tombstones are added to the update chain internally
by prepare transaction rollback and rollback to stable operations. Considering
these tombstones as part of the update obsolete check and trimming the update
chain can cause problem to the threads that are accessing the updates. To avoid
this problem, skip these internally generated globally visible tombstones
as part of the update obsolete checks.

Skipping these tombstones can lead to increase in upates cache usage, but
this can be reclaimed when the page gets evicted or any other follow on update
gets globally visible.