Remove the page image reuse logic as the comparision is not reliable
The page image reuse logic depends on the size and checksum of an image to
be same to reuse, but the checksum function can generate the same checksum
value even when there is a difference in the page images. Reusing an old
image when the new image is modified can lead to corruption.